Added support for building a pdex.dll on windows using a build.bat in the base directory. This creates a win_build folder where all the compilation artifacts go and inside that a FactoryFarming.pdx folder with the game ready to be played on the Windows version of the Playdate Simulator.
This commit does not fix the line-ending problems in the game

The build.bat isn't super fancy. You're welcome to modify and refine this into whatever format you find easiest (You could use a Makefile I'm sure). But this should give you all the info you need for the compiler and linker arguments and other oddities of compiling on Windows.